extract logging config out of /agent dir
Summary:

Test Plan:
