+++ modular-agent-fullstack/backend/src/agent/my_local_agent/route.py
@@ -15,11 +15,11 @@ from rich import print
 
 from .conversation import ConversationManager
 from .examples import get_weather, get_weather_conditions
-from .logging_config import LOGGING_CONFIG
 import traceback
 
 from src.database.db import DatabaseManager, DatabaseUtils
 from src.models import Conversation, Role
+from src.logging_config import LOGGING_CONFIG
 
 
 # Configure logging
